🚧 edit : Bounded `q` in range directly instead of using `qsat`
Use the set method to check if q is in the qlim range and manipulate it accordingly. The problem is that there is a warning saying the set method for the non-dependent property q should not access other properties like qlim. This would probably not cause any problem since the qlim is always modified and assigned before q is changed.

There are two new methods: the first set method is for checking the boundaries of q, and the other is for setting the qlim and updating q which triggers set.q.

NOTE: for now the get.qsat method is commented out but kept for future updates.

TODO:
- validation of any object creation and assignment problems
- keep qsat in mind for future updates probably in the control part
- make dh parameters also accessible